refactor: notification types
