+++ projects/ion/src/lib/notification/notification.component.spec.ts
@@ -1,11 +1,9 @@
 import { EventEmitter } from '@angular/core';
 import { fireEvent, render, screen } from '@testing-library/angular';
 import { StatusType } from '../core/types';
-import {
-  IonNotificationComponent,
-  NotificationProps,
-} from './notification.component';
+import { NotificationProps } from '../core/types/notification';
 import { IonIconModule } from '../icon/icon.module';
+import { IonNotificationComponent } from './notification.component';
 
 const defaultNotification = {
   title: 'Editado',
